+++ src/main/java/com/hidiscuss/backend/controller/ReviewController.java
@@ -3,11 +3,14 @@ package com.hidiscuss.backend.controller;
 import com.hidiscuss.backend.config.SecurityConfig;
 import com.hidiscuss.backend.controller.dto.*;
 import com.hidiscuss.backend.entity.*;
+import com.hidiscuss.backend.service.LiveReviewDiffService;
+import com.hidiscuss.backend.service.ReviewReservationService;
 import com.hidiscuss.backend.service.ReviewService;
 import com.hidiscuss.backend.utils.PageRequest;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
+import lombok.AllArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Sort;
 import org.springframework.http.HttpStatus;
@@ -16,12 +19,17 @@ import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
+import java.util.HashMap;
+import java.util.Map;
 
 @RestController
 @RequestMapping("/review")
+@AllArgsConstructor
 public class ReviewController {
 
     private final ReviewService reviewService;
+    private final LiveReviewDiffService liveReviewDiffService;
+    private final ReviewReservationService reviewReservationService;
 
     public ReviewController(ReviewService reviewService) {
         this.reviewService = reviewService;
@@ -72,7 +80,43 @@ public class ReviewController {
     public Page<ReviewDto> getReviews(@RequestParam("discussionId") Long discussionId, @RequestParam("page") int page) {
         PageRequest pageRequest = new PageRequest(page, Sort.by("createdAt").descending());
         Page<Review> entityPage = reviewService.findAllByDiscussionIdFetch(discussionId, pageRequest.of());
-        Page<ReviewDto> dtoPage = entityPage.map(i -> ReviewDto.fromEntity(i));
-        return dtoPage;
+        return entityPage.map(ReviewDto::fromEntity);
     }
+//------------------------------------------------------------------------
+    @PutMapping("livediff/{diffId}")
+    @ResponseStatus(HttpStatus.CREATED)
+    @Secured(SecurityConfig.DEFAULT_ROLE)
+    @ApiOperation(value = "N분마다 포커싱 되어 있는 파일 업데이트")
+    @ApiResponses({
+            @ApiResponse(code = 200, message = "N분마다 포커싱 되어있는 파일에 대한 liveReviewdiff를 업데이트 한다."),
+            @ApiResponse(code = 400, message = "해당 LiveDiff가 존재하지 않음"),
+            @ApiResponse(code = 500, message = "서버 에러")
+    })
+    public LiveReviewDiffResponseDto updateFocusedDiff(@PathVariable("diffId") Long diffId, @RequestBody UpdateFocusedDiffRequestDto updateFocusedDiffRequestDto, @AuthenticationPrincipal Long userId) {
+        LiveReviewDiff liveReviewDiff = liveReviewDiffService.findByIdAndUpdateByCodeAfter(diffId,updateFocusedDiffRequestDto.codeAfter,userId);
+        return LiveReviewDiffResponseDto.fromEntity(liveReviewDiff);
+    }
+
+    @PutMapping("review/complete/{reviewReservationId}")
+    @Secured(SecurityConfig.DEFAULT_ROLE)
+    @ApiOperation(value = "라이브리뷰 완료")
+    @ApiResponses({
+            @ApiResponse(code = 200, message = "리뷰완료 버튼을 누르거나 한시간이 되면 완료시키는 Api"),
+            @ApiResponse(code = 400, message = "ReviewReservationID가 null 또는 reviewreservation이 존재하지 않음"),
+            @ApiResponse(code = 500, message = "서버 에러")
+    })
+    public Map<String, Long> completeLiveReview(@PathVariable("reviewReservationId") Long reservationId, @RequestBody CompleteLiveReviewRequestDto completeLiveReviewRequestDto) {
+        ReviewReservation reviewReservation = reviewReservationService.findByIdOrNull(reservationId);
+        reviewReservation.setIsdone(Boolean.TRUE);
+        if(reviewReservation.getDiscussion().getState() == DiscussionState.NOT_REVIEWED) {
+            reviewReservation.getDiscussion().setState(DiscussionState.REVIEWING);
+        }
+        reviewReservationService.saveAll(completeLiveReviewRequestDto);
+        HashMap<String, Long> map = new HashMap<>();
+        map.put("discsussionId",reviewReservation.getDiscussion().getId());
+        map.put("reservationId",reviewReservation.getId());
+        return map;
+    }
+
+
 }
