+++ src/main/java/com/hidiscuss/backend/controller/dto/EnterLiveReivewReservationRequestDto.java
@@ -1,0 +1,10 @@
+package com.hidiscuss.backend.controller.dto;
+
+import javax.validation.constraints.NotNull;
+import java.util.List;
+
+public class EnterLiveReivewReservationRequestDto {
+
+    @NotNull
+    public List<Long> discussionCodeIdList;
+}
