+++ src/main/java/com/hidiscuss/backend/controller/dto/UpdateFocusedDiffRequestDto.java
@@ -1,0 +1,9 @@
+package com.hidiscuss.backend.controller.dto;
+
+import javax.validation.constraints.NotNull;
+
+public class UpdateFocusedDiffRequestDto {
+
+    @NotNull
+    public String codeAfter;
+}
