+++ src/main/java/com/hidiscuss/backend/repository/ReviewReservationRepositoryCustom.java
@@ -7,5 +7,5 @@ import java.util.List;
 public interface ReviewReservationRepositoryCustom {
     List<ReviewReservation> findByDiscussionId(Long discussionId);
 
-    List<ReviewReservation> findByDiscussionIdAndUserId(Long userId);
+    List<ReviewReservation> findByUserId(Long userId);
 }
