+++ src/main/java/com/hidiscuss/backend/service/ReviewService.java
@@ -2,15 +2,12 @@ package com.hidiscuss.backend.service;
 
 
 import com.hidiscuss.backend.controller.dto.CreateCommentReviewRequestDto;
-import com.hidiscuss.backend.controller.dto.CreateLiveReviewRequestDto;
 import com.hidiscuss.backend.controller.dto.CreateThreadRequestDto;
-import com.hidiscuss.backend.controller.dto.ReviewDto;
 import com.hidiscuss.backend.entity.*;
 import com.hidiscuss.backend.repository.*;
 import lombok.AllArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -26,7 +23,6 @@ public class ReviewService {
     private final ReviewThreadRepository reviewThreadRepository;
     private final DiscussionRepository discussionRepository;
     private final CommentReviewDiffService commentReviewDiffService;
-    private final LiveReviewDiffService liveReviewDiffService;
 
 
     @Transactional
